+++ packages/kolibri/components/onboarding/TooltipContent.vue
