+++ packages/kolibri/components/onboarding/TooltipTour.vue
@@ -22,7 +22,7 @@
 
   import tippy from 'tippy.js';
   import Vue from 'vue';
-  import { onboardingSteps } from 'kolibri-common/utils/onboardingSteps.js';
+  import { onboardingSteps } from 'kolibri/utils/onboardingSteps.js';
   import TooltipContent from './TooltipContent.vue';
 
   export default {
