+++ packages/kolibri/components/pages/AppBarPage/index.vue
@@ -72,8 +72,8 @@
   import useUser from 'kolibri/composables/useUser';
   import { ref, getCurrentInstance } from 'vue';
   import { useSwipe } from '@vueuse/core';
-  import TooltipTour from 'kolibri.coreVue.components.TooltipTour'; // eslint-disable-line
-  import useTour from 'kolibri.coreVue.composables.useTour'; // eslint-disable-line
+  import TooltipTour from 'kolibri/components/onboarding/TooltipTour'; // eslint-disable-line
+  import useTour from 'kolibri/composables/useTour'; // eslint-disable-line
   import ScrollingHeader from '../ScrollingHeader';
   import AppBar from './internal/AppBar';
   import SideNav from './internal/SideNav';
