+++ packages/kolibri/internal/apiSpec.js
@@ -26,6 +26,8 @@ export default {
   'kolibri/components/error/TechnicalTextBlock': require('kolibri/components/error/TechnicalTextBlock'),
   'kolibri/components/language-switcher/LanguageSwitcherList': require('kolibri/components/language-switcher/LanguageSwitcherList'),
   'kolibri/components/language-switcher/LanguageSwitcherModal': require('kolibri/components/language-switcher/LanguageSwitcherModal'),
+  'kolibri/components/onboarding/TooltipContent': require('kolibri/components/onboarding/TooltipContent'),
+  'kolibri/components/onboarding/TooltipTour': require('kolibri/components/onboarding/TooltipTour'),
   'kolibri/components/pages/AppBarPage': require('kolibri/components/pages/AppBarPage'),
   'kolibri/components/pages/ImmersivePage': require('kolibri/components/pages/ImmersivePage'),
   'kolibri/components/pages/NotificationsRoot': require('kolibri/components/pages/NotificationsRoot'),
@@ -36,6 +38,7 @@ export default {
   'kolibri/composables/useNow': require('kolibri/composables/useNow'),
   'kolibri/composables/useSnackbar': require('kolibri/composables/useSnackbar'),
   'kolibri/composables/useTotalProgress': require('kolibri/composables/useTotalProgress'),
+  'kolibri/composables/useTour': require('kolibri/composables/useTour'),
   'kolibri/composables/useUser': require('kolibri/composables/useUser'),
   'kolibri/composables/useUserSyncStatus': require('kolibri/composables/useUserSyncStatus'),
   'kolibri/constants': require('kolibri/constants'),
@@ -45,6 +48,7 @@ export default {
   'kolibri/styles/themeConfig': require('kolibri/styles/themeConfig'),
   'kolibri/uiText/bytesForHumans': require('kolibri/uiText/bytesForHumans'),
   'kolibri/uiText/commonCoreStrings': require('kolibri/uiText/commonCoreStrings'),
+  'kolibri/uiText/kolibriOnboardingGuideStrings': require('kolibri/uiText/kolibriOnboardingGuideStrings'),
   'kolibri/uiText/licenses': require('kolibri/uiText/licenses'),
   'kolibri/uiText/notificationStrings': require('kolibri/uiText/notificationStrings'),
   'kolibri/urls': require('kolibri/urls'),
@@ -54,6 +58,7 @@ export default {
   'kolibri/utils/browserInfo': require('kolibri/utils/browserInfo'),
   'kolibri/utils/i18n': require('kolibri/utils/i18n'),
   'kolibri/utils/objectSpecs': require('kolibri/utils/objectSpecs'),
+  'kolibri/utils/onboardingSteps': require('kolibri/utils/onboardingSteps'),
   'kolibri/utils/redirectBrowser': require('kolibri/utils/redirectBrowser'),
   'kolibri/utils/scriptLoader': require('kolibri/utils/scriptLoader'),
   'kolibri/utils/serverClock': require('kolibri/utils/serverClock'),
