+++ packages/kolibri/package.json
@@ -33,6 +33,8 @@
     "./components/error/TechnicalTextBlock": "./components/error/TechnicalTextBlock",
     "./components/language-switcher/LanguageSwitcherList": "./components/language-switcher/LanguageSwitcherList",
     "./components/language-switcher/LanguageSwitcherModal": "./components/language-switcher/LanguageSwitcherModal",
+    "./components/onboarding/TooltipContent": "./components/onboarding/TooltipContent",
+    "./components/onboarding/TooltipTour": "./components/onboarding/TooltipTour",
     "./components/pages/AppBarPage": "./components/pages/AppBarPage",
     "./components/pages/ImmersivePage": "./components/pages/ImmersivePage",
     "./components/pages/NotificationsRoot": "./components/pages/NotificationsRoot",
@@ -43,6 +45,7 @@
     "./composables/useNow": "./composables/useNow",
     "./composables/useSnackbar": "./composables/useSnackbar",
     "./composables/useTotalProgress": "./composables/useTotalProgress",
+    "./composables/useTour": "./composables/useTour",
     "./composables/useUser": "./composables/useUser",
     "./composables/useUserSyncStatus": "./composables/useUserSyncStatus",
     "./constants": "./constants",
@@ -52,6 +55,7 @@
     "./styles/themeConfig": "./styles/themeConfig",
     "./uiText/bytesForHumans": "./uiText/bytesForHumans",
     "./uiText/commonCoreStrings": "./uiText/commonCoreStrings",
+    "./uiText/kolibriOnboardingGuideStrings": "./uiText/kolibriOnboardingGuideStrings",
     "./uiText/licenses": "./uiText/licenses",
     "./uiText/notificationStrings": "./uiText/notificationStrings",
     "./urls": "./urls",
@@ -61,6 +65,7 @@
     "./utils/browserInfo": "./utils/browserInfo",
     "./utils/i18n": "./utils/i18n",
     "./utils/objectSpecs": "./utils/objectSpecs",
+    "./utils/onboardingSteps": "./utils/onboardingSteps",
     "./utils/redirectBrowser": "./utils/redirectBrowser",
     "./utils/scriptLoader": "./utils/scriptLoader",
     "./utils/serverClock": "./utils/serverClock",
