+++ packages/kolibri/uiText/kolibriOnboardingGuideStrings.js
